--- a/datafusion/physical-plan/src/sorts/sort.rs
+++ b/datafusion/physical-plan/src/sorts/sort.rs
@@ -50,7 +50,7 @@ use arrow::compute::{concat_batches, lexsort_to_indices, take_arrays, SortColumn
 use arrow::datatypes::{DataType, SchemaRef};
 use arrow::row::{RowConverter, Rows, SortField};
 use datafusion_common::{
-    exec_datafusion_err, internal_datafusion_err, internal_err, Result,
+    config_err, exec_datafusion_err, internal_datafusion_err, internal_err, Result,
 };
 use datafusion_execution::disk_manager::RefCountedTempFile;
 use datafusion_execution::memory_pool::{MemoryConsumer, MemoryReservation};
@@ -60,7 +60,7 @@ use datafusion_physical_expr::LexOrdering;
 use datafusion_physical_expr_common::sort_expr::LexRequirement;
 
 use futures::{StreamExt, TryStreamExt};
-use log::{debug, trace};
+use log::trace;
 
 struct ExternalSorterMetrics {
     /// metrics
@@ -1338,6 +1338,16 @@ impl ExecutionPlan for SortExec {
                 )))
             }
             (false, None) => {
+                // Validate `ExecutionOptions::sort_max_spill_merge_degree` (must be >= 2)
+                let spill_max_merge_degree =
+                    execution_options.sort_max_spill_merge_degree;
+                if spill_max_merge_degree < 2 {
+                    return config_err!(
+                        "sort_max_spill_merge_degree must be >= 2 in order to continue external sorting, but got {}",
+                        spill_max_merge_degree
+                    );
+                }
+
                 let mut sorter = ExternalSorter::new(
                     partition,
                     input.schema(),
@@ -1345,7 +1355,7 @@ impl ExecutionPlan for SortExec {
                     context.session_config().batch_size(),
                     execution_options.sort_spill_reservation_bytes,
                     execution_options.sort_in_place_threshold_bytes,
-                    execution_options.sort_max_spill_merge_degree,
+                    spill_max_merge_degree,
                     &self.metrics_set,
                     context.runtime_env(),
                 )?;