… Cargo.lock
… datafusion/core/tests/fuzz_cases/sort_fuzz.rs
… datafusion/core/tests/memory_limit/mod.rs
… datafusion/sqllogictest/test_files/information_schema.slt
… docs/source/user-guide/configs.md