@rluvaton 2025-04-07
BTW, row_hash uses the sort preserving merge stream as well and has similar problem, I think this should be a solution outside the sort exec