@2010YOUY01 2025-04-07
> Also, to have a fully working larger than memory sort, you need to spill in
> 
> https://github.com/apache/datafusion/blob/362fcdfc7b9e00cb6126a0cbc41c9abb2637c563/datafusion/physical-plan/src/sorts/builder.rs#L74
> 
> In case the memory reservation is failing

Could you elaborate? I don't get it.